+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## FUTURE
 
 * Add a bluetti-discovery command that can collect information from unsupported Bluetti devices
+* Out-of-range DC input current values are no longer reported
 
 ## 0.11.1